valnum, numformatter moved to lib-wx-init
