+++ libraries/lib-wx-init/numformatter.cpp
