+++ libraries/lib-wx-init/numformatter.h
