+++ libraries/lib-wx-init/valnum.cpp
