+++ src/CMakeLists.txt
@@ -1043,10 +1043,6 @@ list( APPEND SOURCES
       widgets/Warning.h
       widgets/auStaticText.cpp
       widgets/auStaticText.h
-      widgets/numformatter.cpp
-      widgets/numformatter.h
-      widgets/valnum.cpp
-      widgets/valnum.h
 
       # Update version
       $<$<BOOL:${${_OPT}has_updates_check}>:
