+++ src/commands/Demo.cpp
@@ -28,7 +28,7 @@ parameters.  It is for development purposes.
 #include "SettingsVisitor.h"
 #include "ShuttleGui.h"
 #include "AudacityMessageBox.h"
-#include "../widgets/valnum.h"
+#include "valnum.h"
 #include "../commands/CommandContext.h"
 
 const ComponentInterfaceSymbol DemoCommand::Symbol
