+++ src/effects/Amplify.cpp
@@ -37,7 +37,7 @@
 
 #include "ShuttleGui.h"
 #include "WaveTrack.h"
-#include "../widgets/valnum.h"
+#include "valnum.h"
 
 
 enum
