+++ src/effects/AutoDuck.cpp
@@ -29,7 +29,7 @@
 #include "Prefs.h"
 #include "ShuttleGui.h"
 #include "Theme.h"
-#include "../widgets/valnum.h"
+#include "valnum.h"
 
 #include "WaveTrack.h"
 #include "AudacityMessageBox.h"
