+++ src/effects/Distortion.cpp
@@ -43,7 +43,7 @@
 
 #include "Prefs.h"
 #include "ShuttleGui.h"
-#include "../widgets/valnum.h"
+#include "valnum.h"
 
 const EnumValueSymbol EffectDistortion::kTableTypeStrings[nTableTypes] =
 {
