+++ src/effects/DtmfGen.cpp
@@ -26,7 +26,7 @@
 #include "Prefs.h"
 #include "ShuttleGui.h"
 #include "../widgets/NumericTextCtrl.h"
-#include "../widgets/valnum.h"
+#include "valnum.h"
 
 
 namespace {
