+++ src/effects/Noise.cpp
@@ -24,7 +24,7 @@
 
 #include "Prefs.h"
 #include "ShuttleGui.h"
-#include "../widgets/valnum.h"
+#include "valnum.h"
 #include "../widgets/NumericTextCtrl.h"
 
 const EnumValueSymbol EffectNoise::kTypeStrings[nTypes] =
