+++ src/effects/Normalize.cpp
@@ -27,7 +27,7 @@
 #include "../ProjectFileManager.h"
 #include "ShuttleGui.h"
 #include "WaveTrack.h"
-#include "../widgets/valnum.h"
+#include "valnum.h"
 #include "ProgressDialog.h"
 
 const EffectParameterMethods& EffectNormalize::Parameters() const
