+++ src/effects/Paulstretch.cpp
@@ -25,7 +25,7 @@
 
 #include "ShuttleGui.h"
 #include "FFT.h"
-#include "../widgets/valnum.h"
+#include "valnum.h"
 #include "AudacityMessageBox.h"
 #include "Prefs.h"
