+++ src/effects/Phaser.cpp
@@ -26,7 +26,7 @@
 #include <wx/slider.h>
 
 #include "ShuttleGui.h"
-#include "../widgets/valnum.h"
+#include "valnum.h"
 #include <wx/weakref.h>
 
 const EffectParameterMethods& EffectPhaser::Parameters() const
