+++ src/effects/Repeat.cpp
@@ -31,7 +31,7 @@
 #include "WaveTrack.h"
 #include "WaveClip.h"
 #include "../widgets/NumericTextCtrl.h"
-#include "../widgets/valnum.h"
+#include "valnum.h"
 
 #include "LoadEffects.h"
