+++ src/effects/Reverb.cpp
@@ -24,7 +24,7 @@
 
 #include "Prefs.h"
 #include "ShuttleGui.h"
-#include "../widgets/valnum.h"
+#include "valnum.h"
 
 #include "Reverb_libSoX.h"
