+++ src/effects/ToneGen.cpp
@@ -29,7 +29,7 @@ frequency changes smoothly during the tone.
 #include "Project.h"
 #include "ProjectRate.h"
 #include "ShuttleGui.h"
-#include "../widgets/valnum.h"
+#include "valnum.h"
 #include "../widgets/NumericTextCtrl.h"
 
 const EnumValueSymbol EffectToneGen::kInterStrings[nInterpolations] =
