+++ src/effects/TruncSilence.cpp
@@ -32,7 +32,7 @@
 #include "ShuttleGui.h"
 #include "SyncLock.h"
 #include "WaveTrack.h"
-#include "../widgets/valnum.h"
+#include "valnum.h"
 #include "AudacityMessageBox.h"
 
 class Enums {
