+++ src/effects/Wahwah.cpp
@@ -29,7 +29,7 @@
 #include <wx/weakref.h>
 
 #include "ShuttleGui.h"
-#include "../widgets/valnum.h"
+#include "valnum.h"
 
 const EffectParameterMethods& EffectWahwah::Parameters() const
 {
