+++ src/effects/nyquist/Nyquist.cpp
@@ -70,7 +70,7 @@ effects from this one class.
 #include "ViewInfo.h"
 #include "WaveClip.h"
 #include "WaveTrack.h"
-#include "../../widgets/valnum.h"
+#include "valnum.h"
 #include "AudacityMessageBox.h"
 #include "Prefs.h"
 #include "wxFileNameWrapper.h"
