+++ src/effects/vamp/VampEffect.cpp
@@ -32,7 +32,7 @@
 
 
 #include "ShuttleGui.h"
-#include "../../widgets/valnum.h"
+#include "valnum.h"
 #include "AudacityMessageBox.h"
 
 #include "../../LabelTrack.h"
